lib/igt_vc4: Rework igt_vc4_get_param()
Instead of returning the error code of the DRM_IOCTL_VC4_GET_PARAM ioctl,
make igt_vc4_get_param() return the value of the parameter, considering
that the current return is not being used and the value of the parameter
is being used. If an invalid parameter is passed to the ioctl, return 0.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>