lib/igt_v3d: Rework igt_v3d_get_param()
Instead of assuming that igt_v3d_get_param() will succeed, allow the
ioctl to fail and return 0. This way, it will be possible to check if the
kernel has the capabilities needed by the test by checking if the return
of the igt_v3d_get_param() function is different than zero.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>