i915/gem_create: Ensure newly created handles are not reusing same offset
We need to ensure second group of spinners won't stuck on vma eviction
and they will be scheduled for the execution. We need to propose new
offsets to avoid situation spinner batch and object dependency handle
offset will overlap. Since reloc allocator is tracking offsets for
handles we need to explicitly release it.

Fixes: https://gitlab.freedesktop.org/drm/intel/-/issues/7679

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>