i915/gem_(linear, tiled)_blits: Randomise buffer contents
Currently, we use an incrementing value for the buffer contents,
starting the next buffer from the final value of the last. This means
that the value of corresponding dwords between two buffers is offset
by a single bit. In order to differentiate between an error in copying
between two buffers from single bit memory errors, we need to randomise
the offset between those two buffers.

Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Signed-off-by: Nirmoy Das <nirmoy.das@intel.com>
Reviewed-by: Nirmoy Das <nirmoy.das@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>