i915/gem_ccs: Change allocator type and enforce softpinning
For all subtests which uses simple allocator change them to reloc.
There's nothing wrong with simple, but when it is freeing offset
it might be used again in next allocation. From the execution point
of view it leads to stalls. We don't want that here, especially in
ctrl-surf-copy* subtests, where we join restoring ccs data for the
surface and block-copy immediate after that.

Ensure we use softpinning for all subtests for the case where kernel
still supports relocations and get_*_ahnd() helpers would return
invalid allocator handle.

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Karolina Stolarek <karolina.stolarek@intel.com>
Reviewed-by: Karolina Stolarek <karolina.stolarek@intel.com>