tests/vkms: Create a testlist to the vkms DRM driver
The vkms DRM driver is a software-only model of a KMS driver and its
development is highly based on the current status of the IGT tests. So,
create a list of the tests that are supported by the vkms DRM driver, in
order to have a baseline of the tests that should pass when using vkms.

This way, it will easier to track regressions on the vkms DRM driver, as
the testlist for the driver will be publicly available.

Acked-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Maíra Canal <mcanal@igalia.com>