lib/intel_tiling_info: Rename blt_cmd_info to intel_cmds_info
The struct stores information on more than just one blitter cmd,
so the current name is slightly confusing. Also, we might want
to store information on other copy commands, like the one provided
by render. Change it to intel_cmds_info to emphasize that it
describes available commands on the platform. Change the field
name in intel_device_info and update related definitions to refer
to commands, not tiling.

Signed-off-by: Karolina Stolarek <karolina.stolarek@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>