lib/i915_blt: Add checks for command properties
Add a predicate that checks if a command has a specific property.
Introduce helpers that check if a platform uses an extended
version of the block-copy command or supports compression. Add
a getter for blt_cmd_info to intel_cmds_info library. Group
functions together in i915_blt header file.

Signed-off-by: Karolina Stolarek <karolina.stolarek@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>