lib/igt_kms: handle spurious HPDs - IGT part
Some panels generate long HPDs during CI execution steps even
while connected to the system and cause unexpected CI execution
failures. In environments like CI, we don't expect to disconnect
the panels in the middle of the CI execution.

There are two parts to handle this case - display driver and IGT

1. In the display driver, based on a flag, long HPDs are
ignored. This flag can be set/unset using debugfs on systems
where such panels are connected.

2. In IGT, add provision to set ignore long HPD debugfs entry
on the driver and also set Force "on" the active connectors.
With force on, the connector's detect sequences will not get
initiated.

This patch handles the IGT part.

An enviroment variable "IGT_KMS_IGNORE_HPD" is added to
differentiate systems which require such spurious HPD handling.
If this variable is on, ignore long HPD debugs entry is set and
active connectors are set to force "on" state.

Many thanks to "Imre Deak" for the suggestions and support

v2: Minor updates in the comments.

Cc: Imre Deak <imre.deak@intel.com>
Signed-off-by: Vinod Govindapillai <vinod.govindapillai@intel.com>
Reviewed-by: Imre Deak <imre.deak@intel.com>