tests/kms_plane_scaling: Reorg max-source-size subtest
The i915-max-source-size test has a restriction that it will
only run on displays with a resolution of 3840x2160. Initially,
the test was written to be executed with a single output,
so on a multi-display setup, even if there was a display with a
resolution of 3840x2160, the test would skip if the first
output didn't support that resolution.

To address this issue, the test has been modified so that it
can run on each valid connector per pipe.

v2: -remove "i915-max-source-size" from dynamic subtest name (Karthik)

Signed-off-by: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>