xe/xe_test_config.json: cleanup the field struct
Originally, the JSON file was conceived to also validate fields
and only allow certain values when others are present.

It turns that it doesn't make much sense to do such validations,
as it makes harder to change the hierarchy. So, remove the
parts of the fields struct that doesn't contain "is_field".

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Acked-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>