intel_gpu_top: Use actual period when calculating client busyness
On a slow machine, or with many processes and/or file descriptors to
parse, the period of the scanning loop can drift significantly from the
assumed value. This results in artificially inflated client busyness
percentages.

To alleviate the issue take some real timestamps and use actual elapsed
time when calculating relative busyness.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>