tests/kms_plane_scaling: Fix out-of-bound array access
With this fix we are solving 2 issues. Firstly, the
planes_scaling_combo() tests were leaving one scaler assigned
after running sub-test with two consecutive planes because
one scaler was getting reused in the next run. So with this fix
scaler is not reused since we won't have any common plane.

Secondly, when k == n_planes - 1, we were trying to access
planes[n_planes], which led to array out of bounds error. So, with
this fix, this issue is fixed too.

v2: -fix condition if n_planes is not even (JP)

Suggested-by: Luca Coelho <luciano.coelho@intel.com>
Suggested-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Reported-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>
Reviewed-by: Luca Coelho <luciano.coelho@intel.com>