meson.build: place gtk-doc dependencies check at the right place
The requirement for having gtk-doc is specific to
docs/reference/igt-gpu-tools.

Moved the code to be there. This lets build testplan docs
without needing gtk-doc (nor having the IGT executables).

It should be noticed that testplan indirectly depends on
build_tests, for one reason only: the tests subdir is only
included if build_tests is true. Without it, the xe_test_config
variable will be undefined, as it is declared inside a tests
subdir.

To ensure that the variable is declared we need to add an extra
test for the testplan subdir.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>