tests/kms_color: Use legacy LUT for chopping off the lsbs
Trying to use an interpolated gamma mode (as is the case currently
on eg. icl/tgl) for chopping off the lsbs doesn't really work. Use
the legacy LUT instead. Also get rid of the degamma LUT as that
can't even be enabled alognside the legacy LUT on split gamma i915
platforms (ivb-skl).

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Reviewed-by: Uma Shankar <uma.shankar@intel.com>