testplan/meson.build: only run igt_runner if it is compiled
While the runner is built by default, if jsonc is not found, it
won't be build. This will cause the documentation validation test
to cause the build to fail.

Prevent that by adding an extra check if jsonc is found, as this
is the same condition that it is used to build igt_runner.

While here, also ensure that igt_doc.py will depend on having the
igt_runner already compiled, when the --check option is used.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>