lib/xe_query: fix vram_memory() and vram_if_possible()
vram_memory() is meant to return VRAM if supported. If we call this on
igpu it should return zero. There is at least one user relying on this
in xe_mmap@vram, where we want to skip if the mask here is zero.

vram_if_possible() should account for the per GT VRAM when returning
the mask. Currently it just returns the mask for all VRAM instances.

v2 (Niranjana):
  - Squash into a single patch
  - Fold in the differences from the version by Matt Brost & Zbigniew

Reported-by: Gwan-gyeong Mun <gwan-gyeong.mun@intel.com>
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>