tests/vkms: Add tests that perform operations on the primary plane
Commits 4dee3c4b32a6 ("drm/vkms: remove the need for the primary plane
to be visible") and 4a1b5d183273 ("drm/vkms: allow the primary plane to
be positioned") allowed vkms to have a non-visible primary plane and
to position the primary plane in such a way that doesn't cover the
entire CRTC.

Those commits ended up making it possible to run tests that perform
operations on the primary plane. So, add tests that perform operations
on the primary plane.

Signed-off-by: Maíra Canal <mcanal@igalia.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>