tests/xe_vm: make sure test can finish within time
the BAT takes a lot longer (> 48s) than expected (~5s). This seems
due to the system is too busy to wake up the main thread which is in
sleep.

So, change the test to run in a timeout instead of relying on main
thread.

on another hand, also change the logic to make sure all threads start
together.

Signed-off-by: Bruce Chang <yu.bruce.chang@intel.com>
Reviewed-by: Niranjana Vishwanathapura <niranjana.vishwanathapura@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>