meson_options.txt: disable Sphinx by default
Building IGT with Sphinx enabled and rst2pdf installed takes:

	$ touch tests/xe/xe_compute.c
	$ time ninja -C build

	...

        real    0m10.545s
        user    0m13.778s
        sys     0m0.724s

With Sphinx disabled, the same build is a lot faster:

	$ touch tests/xe/xe_compute.c
	$ time ninja -C build

	...

	real	0m1.746s
	user	0m1.574s
	sys	0m0.171s

As Sphinx and PDF outputs aren't something that every single
IGT developer would need, let's speedup the build by disabling
it by default.

Test plan and test documentation validation will still run,
as this is a necessary step to validate if the tests documentation
were not broken on a patch, but this will save some time during
normal builds.

Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>