tools/intel-gfx-fw-info: Make it compatible with HuC via GSC
When HuC is loaded via GSC, the firmware format is different and there
is no information for kernel to parse except the version, that is in
a different location.  Check for the magic field as the first dword and
parse the blob differently based on that.

Tested with
https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/tree/i915/dg2_huc_gsc.bin?id=8f86b5ab3e051170ea240fc409d457e16e24bc21,
with output "version: 7.10.3" as expected. Also checked with a couple
of GuC firmware blobs and CSS-based HuC firmware blobs to guarantee
this doesn't regress.

v2: Add explicit ver_hi and ver_lo field to better visualize dump (Bala)

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>