scripts/test_list.py: expand testlist fields
We're about to add support for i915 testlists. On those,
it is possible to have functionalities mapped to different
buckets, as the same test could be used to test multiple
parts of the driver/hardware.

So, for instance, the subtest igt@prime_vgem@basic-fence-mmap
actually tests two features:
	- gtt
	- prime

A description of such test will then be mapped as:

	* SUBTEST: basic-fence-mmap
	* Description: Examine GTT access path fencing.
	* Feature: gtt, prime

In order to be able to generate per-feature testlists, we
need to expand the Feature field, using the separator field
(currently a comma) to expand the test lists.

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>