i915_test_config.json: exclude gem_concurrent_blit.c
This is actually a subset of gem_concurrent_all.c, which can't
be described, as it containd 400k+ tests on it. It also has an
insane number of tests:

	$ ./build/tests/gem_concurrent_all --list |wc -l
	428400
	$ ./build/tests/gem_concurrent_blit --list |wc -l
	57120

It makes no sense to document all of them. So, exclude it as
well, as otherwise adding a logic to check for missing documentation
on i915 will complain with:

	...
	Warning: Missing documentation for igt@gem_concurrent_blit@4KiB-tiny-wc-blt-read-read-bcs-forked
	Warning: Missing documentation for igt@gem_concurrent_blit@4KiB-tiny-wc-blt-read-read-bcs-interruptible
	Warning: Missing documentation for igt@gem_concurrent_blit@4KiB-tiny-wc-blt-read-read-rcs
	...

Reviewed-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>