lib/intel_bb: Enable custom engine support for xe
Currently the 'ctx' field in batch buffer creation is interpreted as
a vm id for xe. In i915 it is interpreted as a context id. Since a xe
engine more closely resembles an i915 context, change the current
behaviour and interpret the 'ctx' fied as an xe engine id. This also
allows us to use the compute engine on xe, which currently is not
possible, due to reliance on legacy i915 flags.

v2:
 - don't destroy user provided engine in 'intel_bb_destroy' (Zbigniew)
 - destroy internally created engine before creating a new one
v3:
 - introduce 'vm' parameter to be able to provide probler vm_id for
   custom xe engine (Dominik, Zbigniew)

Signed-off-by: Christoph Manszewski <christoph.manszewski@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>