lib/intel_batchbuffer: Use COMPUTE instead RENDER engine for Xe on PVC
For gpgpu fill currently we select I915_EXEC_RENDER which is correct
as pipeline selection allows to submit compute job on it. However
newer platforms like PVC have no render engine so replacing to compute
engine is necessary. At the moment this is workaround - more rework
in intel-bb is necessary to better support both - i915 and xe.

Signed-off-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Reviewed-by: Christoph Manszewski <christoph.manszewski@intel.com>