tests/xe/intel_bb: remove full-batch
Currently this generates CAT errors due to over-fetch going beyond the
boundaries of the GTT. However we also don't use scratch pages by
default in Xe (hopeful plan is to get rid them completely), so no
address with full-batch will be valid if HW over-fetches into the next
page, unless we either inflate the batch size with padding, which
defeats the purpose of the test, or enable scratch pages, which is not
something we really want moving forward. With that rather just remove
full-batch on Xe.

Bspec: 60223
Closes: https://gitlab.freedesktop.org/drm/xe/kernel/-/issues/262
Signed-off-by: Matthew Auld <matthew.auld@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>